requirements.txt
python-telegram-bot==21.3
httpx==0.27.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
playwright==1.44.0
